+++ build.sc
@@ -2,9 +2,12 @@ import mill._, scalalib._
 
 val spinalVersion = "1.7.3"
 
-object mylib extends SbtModule {
-  def scalaVersion = "2.12.14"
+object projectname extends SbtModule {
+  def scalaVersion = "2.12.16"
   override def millSourcePath = os.pwd
+  def sources = T.sources(
+    millSourcePath / "hw" / "spinal"
+  )
   def ivyDeps = Agg(
     ivy"com.github.spinalhdl::spinalhdl-core:$spinalVersion",
     ivy"com.github.spinalhdl::spinalhdl-lib:$spinalVersion"
